+++ CHANGELOG.md
@@ -1,3 +1,12 @@
+#### [1.0.0] - 2016-19-01
+* UIColor+HexColor - moved out of pod
+* new documentation for `VBPieChart` class
+* fixed issues with passing wrong data
+* new `VBPieChartAnimationDefault`
+* new `insertChartValue:atIndex:`
+* reverted `- (NSArray *) chartValues;`
+* fixed accent vector
+* some more Unit Tests
 
 #### [0.4.1] - 2016-18-01
 * fixed crash for empty `chartValues`